Local training now works

+++ launch_script.py
@@ -1,21 +1,21 @@
 # this scripts installs necessary requirements and launches main program
 import subprocess
 import os
-import sys
 import importlib.util
+import sys
 import urllib.request
 from shutil import which
 
+# ****************************************************************************
+# *                                   UTIL                                   *
+# ****************************************************************************
+
 python = sys.executable
-git = os.environ.get('GIT', "git")
-index_url = os.environ.get('INDEX_URL', "")
 platform = sys.platform
 
 def prRed(skk): print(f"\033[91m{skk}\033[00m") 
 def prGreen(skk): print(f"\033[92m{skk}\033[00m")
 def prYellow(skk): print(f"\033[93m{skk}\033[00m")
-
-prGreen('Starting launch script...')
 
 def run(command, desc=None, errdesc=None, custom_env=None, live=False):
     if desc is not None:
@@ -54,33 +54,87 @@ def is_installed(package):
 
 
 def run_pip(args, desc=None):
+    index_url = os.environ.get('INDEX_URL', "")
     index_url_line = f' --index-url {index_url}' if index_url != '' else ''
-    return run(f'"{python}" -m pip {args} --prefer-binary{index_url_line}', desc=f"Installing {desc}", errdesc=f"Couldn't install {desc}")
+    return run(f'"{sys.executable}" -m pip {args} --prefer-binary{index_url_line}', desc=f"Installing {desc}", errdesc=f"Couldn't install {desc}")
 
 
 
+
+
+# ****************************************************************************
+# *                                  CONFIG                                  *
+# ****************************************************************************
+
+SKIP_INSTALL = False
+
+torch_command = "pip install torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu117"
+
+
+main_script_path = "main_gui.py"
+pre_torch_packages = []
+post_torch_packages = [
+                        "black",
+                        "diffusers",
+                        "tqdm",
+                        "PySimpleGUI",
+                        "torchaudio",
+                        "pydub",
+                        "ema_pytorch",
+                        "pytorch_lightning",
+                        "pygame",
+                        "soundfile",
+                        "prefigure",
+                        "wandb",
+                        "einops",
+                        "matplotlib",
+                        "pandas",
+                        "librosa",
+                        "tqdm"
+                        ]
+
+
+
+prGreen('Starting launch script...')
 
 if __name__ == "__main__":
-    # env setup check
-    if not is_installed("torch"):
-        torch_command = os.environ.get('TORCH_COMMAND', "pip install torch==1.13.1+cu117 --extra-index-url https://download.pytorch.org/whl/cu117")
-        run(f'"{python}" -m {torch_command}', "Installing torch.", "Couldn't install torch", live=True)
-        run_pip(f"install .", "requirements")
+
+    # INSTALL
+
+    if not SKIP_INSTALL:
+
+        # pre torch packages
+        if pre_torch_packages:
+            for package in pre_torch_packages:
+                if not is_installed(package):
+                    run_pip(f"install {package}", package)
 
 
-    if which('ffmpeg') is None:
-        if platform == 'win32':
-            prYellow('No FFMPEG detected! Opening installer..')
-            urllib.request.urlretrieve("https://github.com/icedterminal/ffmpeg-installer/releases/download/5.1.0.20220727/FFmpeg_Essentials.msi", "ffmpeg_installer.msi")
-            subprocess.run("ffmpeg_installer.msi", shell=True)
-            os.remove("ffmpeg_installer.msi")
+        # TORCH INSTALL
+        if not is_installed("torch") and torch_command is not None:
+            run(f'"{sys.executable}" -m {torch_command}', "Installing torch.", "Couldn't install torch", live=True)
+        
+        # post torch packages
+        if post_torch_packages:
+            for package in post_torch_packages:
+                if not is_installed(package):
+                    run_pip(f"install {package}", package)
 
-        elif platform == 'linux':
-            prRed('No FFMPEG detected! Please update pkgs and run "sudo apt install ffmpeg" to install before using the GUI!')
-            exit()
 
-        elif platform == 'darwin':
-            prRed('No FFMPEG detected! Please use homebrew "brew install ffmpeg" to install before using the GUI!')
-            exit()
-
-    import main_gui
+        
+        if which('ffmpeg') is None:
+                if platform == 'win32':
+                    prYellow('No FFMPEG detected! Opening installer..')
+                    urllib.request.urlretrieve("https://github.com/icedterminal/ffmpeg-installer/releases/download/5.1.0.20220727/FFmpeg_Essentials.msi", "ffmpeg_installer.msi")
+                    subprocess.run("ffmpeg_installer.msi", shell=True)
+                    os.remove("ffmpeg_installer.msi")
+    
+                elif platform == 'linux':
+                    prRed('No FFMPEG detected! Please update pkgs and run "sudo apt install ffmpeg" to install before using the GUI!')
+                    exit()
+    
+                elif platform == 'darwin':
+                    prRed('No FFMPEG detected! Please use homebrew "brew install ffmpeg" to install before using the GUI!')
+                    exit()
+    # LAUNCH
+    run(f'"{sys.executable}" {main_script_path}', "Starting main script..", "Couldn't start main script!", live=True)
